+++ NEWS.md
@@ -11,6 +11,7 @@
 * Fix `google_analytics_account_list()` if only one view available (#52, #41)
 * Add `antiSampleWorked` TRUE/FALSE attribute to tables when anti-sampling attempted
 * Add a `slow_fetch` flag to `google_analytics_4` that will avoid batching for big complicated fetchess
+* Add `ga_users_list` for listing users on account, webProperty or View level
 
 ## Fixed